+++ BadWolfProduction/src/main/java/com/nowscas/BadWolfProduction/repos/MainPagePostRepo.java
@@ -6,5 +6,5 @@ import org.springframework.data.repository.CrudRepository;
 /**
  * Интерфейс для работы с записями постов в БД.
  */
-public interface MainPagePostRepo extends CrudRepository<MainPagePost, Integer> {
+public interface MainPagePostRepo extends CrudRepository<MainPagePost, Long> {
 }
